Add mapbox-gl-draw styles for button icons, create MapPolygonFilter component
