+++ atd-vzv/package.json
@@ -14,6 +14,7 @@
     "@fortawesome/fontawesome-svg-core": "^1.2.26",
     "@fortawesome/free-solid-svg-icons": "^5.12.0",
     "@fortawesome/react-fontawesome": "^0.1.8",
+    "@mapbox/mapbox-gl-draw": "^1.1.2",
     "@material-ui/core": "^4.7.1",
     "@turf/difference": "6.0.1",
     "aphrodite": "^2.4.0",
